Passing VectorNormals tests
A few issues addressed:

- There were what looked a lot like finite-precision
errors happening in the normals funcs when using
cartesian_positions as inputs. Fixed by now using
float64 when performing that arithmetic

- Signs were getting flipped in the octencoded
vertex normal values. This went away when I implemented
the TODO item in oct_encode

- Tests now comparing vertex normals with an allowable
error of tolerance to account for accuracy lost in
the octencoding/decoding

+++ quantized_mesh_encoder/util_cy.pyx
@@ -78,8 +78,8 @@ def ritter_second_pass(
 #         vertex_normals[pos] += face_norm
 def add_vertex_normals(
     np.ndarray[np.uint32_t, ndim=2] indices,
-    np.ndarray[np.float32_t, ndim=2] normals,
-    np.ndarray[np.float32_t, ndim=2] out):
+    np.ndarray[np.float64_t, ndim=2] normals,
+    np.ndarray[np.float64_t, ndim=2] out):
 
     cdef long long indices_length = indices.shape[0]
     cdef Py_ssize_t i, j, k
